[refac] notes persist function into persists
